Variety of parameter tweaks and bug fixes, decent performance
Updated lane line filter to take the average of the inputs
that pass through the rejection filter, and update the
parameters for the hough transform to be more lenient.

The solid white and yellow line example are decent, the challenge
video performance is poor when the gradient between the road
surface and yellow paint reduces.